ml/scripts/requirements.txt
transformers
torch
datasets
peft
bitsandbytes
accelerate
scipy
evaluate
pyyaml
tqdm
pandas
numpy